+++ src/Translatable/TranslatableServiceProvider.php
@@ -10,7 +10,7 @@ class TranslatableServiceProvider extends ServiceProvider
     {
         $this->publishes([
             __DIR__.'/../config/translatable.php' => config_path('translatable.php'),
-        ]);
+        ], 'translatable');
     }
 
     /**