+++ src/config/translatable.php
@@ -58,6 +58,18 @@ return [
 
     /*
     |--------------------------------------------------------------------------
+    | Use fallback per property
+    |--------------------------------------------------------------------------
+    |
+    | The property fallback feature will return the translated value of
+    | the fallback locale if the property is empty for the selected
+    | locale. Note that 'use_fallback' must be enabled.
+    |
+     */
+    'use_property_fallback' => true,
+
+    /*
+    |--------------------------------------------------------------------------
     | Fallback Locale
     |--------------------------------------------------------------------------
     |
@@ -91,4 +103,14 @@ return [
     */
     'locale_key' => 'locale',
 
+    /*
+    |--------------------------------------------------------------------------
+    | Always load translations when converting to array
+    |--------------------------------------------------------------------------
+    | Setting this to false will have a performance improvement but will
+    | not return the translations when using toArray(), unless the
+    | translations relationship is already loaded.
+    |
+     */
+    'to_array_always_loads_translations' => true,
 ];